modified index.js and categoriesSlice.js
