+++ src/redux/categories/categoriesSlice.js
@@ -2,15 +2,17 @@ import { createSlice } from '@reduxjs/toolkit';
 
 const initialState = {
   categories: [],
-  status: 'Under construction',
 };
 
 const categoriesSlice = createSlice({
   name: 'categories',
   initialState,
   reducers: {
-    // No additional reducers for categories slice
+    setStatus: (state, action) => {
+      state.status = action.payload;
+    },
   },
 });
 
+export const { setStatus } = categoriesSlice.actions;
 export default categoriesSlice.reducer;
